% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetRes.R
\name{NetRes}
\alias{NetRes}
\title{NetRes NetRes class definition}
\value{
Nothing
}
\description{
NetRes NetRes class definition

NetRes NetRes class definition
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ensemble}}{The final ensemble as inferred after some iterations}

\item{\code{latent.space}}{The list of latent spaces after a given iteration}

\item{\code{train.data}}{The initial training data}

\item{\code{latent.data}}{The true data for the latent space (if provided)}

\item{\code{true.graph}}{The true graph (if any)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NetRes$new()}}
\item \href{#method-assess}{\code{NetRes$assess()}}
\item \href{#method-plot}{\code{NetRes$plot()}}
\item \href{#method-clone}{\code{NetRes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor that fits a NetRes object from a data fram
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetRes$new(
  dframe,
  true.graph = NULL,
  nIter,
  nBoot = 50,
  algorithm = "tabu",
  algorithm.args,
  lvPrefix = "^U\\\\_",
  mode = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dframe}}{Training data frame, optionally with true latent variables (these will be used for performance reporting)}

\item{\code{nIter}}{Number of iterations (NULL to autodetect - not yet supported)}

\item{\code{nBoot}}{Number of bootstraps (default = 50)}

\item{\code{algorithm}}{Algorithm to use (default = 'tabu')}

\item{\code{algorithm.args}}{Arguments to the algorithm (must be a list)}

\item{\code{lvPrefix}}{Prefix of variables to be treated as latent (if any).  If not present, defautls to "U\\_".  If no latent vars found, 
no performance assessment of latent discovery will (or indeed can) be made, but whatever latent space is discovered will be duly reported}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A NetRes object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assess"></a>}}
\if{latex}{\out{\hypertarget{method-assess}{}}}
\subsection{Method \code{assess()}}{
assess Assess the inferred ensemble against the true graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetRes$assess(true.graph = self$true.graph, lvPrefix = "^U\\\\_")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true.graph}}{The true graph to use; defaults to the one provided at initialization (if any)}

\item{\code{lvPrefix}}{The latent variable-identifying regular expression, as elsewhere; defaults to "^U\\_"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot The function to plot (some) networks in the ensemble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetRes$plot(networks = NULL, lvPrefix = "^U\\\\_")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{networks}}{Indexes into the ensemble}

\item{\code{lvPrefix}}{The latent variable prefix (default = 'U_'; use perl regexp)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetRes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
