% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFNetwork.R
\name{SFNetwork}
\alias{SFNetwork}
\title{SFNetwork SFNetwork class definition}
\description{
SFNetwork SFNetwork class definition

SFNetwork SFNetwork class definition
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dag}}{The generated directed acyclic graph}

\item{\code{vRank}}{vertex topological rank}

\item{\code{outDegree}}{out-degree of vertices based on direct descendants}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SFNetwork$new()}}
\item \href{#method-print}{\code{SFNetwork$print()}}
\item \href{#method-plot}{\code{SFNetwork$plot()}}
\item \href{#method-generateData}{\code{SFNetwork$generateData()}}
\item \href{#method-clone}{\code{SFNetwork$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Generate a DAG using Barabasi's game and generate matching data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFNetwork$new(
  numVertices = 20,
  topology = "power",
  numNeighbors = 4,
  nParentOnly = 4
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numVertices}}{number of vertices in the DAG (default=20)}

\item{\code{topology}}{power/star}

\item{\code{numNeighbors}}{number of neighbors of each vertex, on average (i.e., node denisty, default=4)}

\item{\code{nParentOnly}}{how many nodes to make parent-only hubs by removing their high order but low out-degree parents (default=4)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `SFNetwork` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
prints the details of the DAG
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFNetwork$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plots the DAG
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFNetwork$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateData"></a>}}
\if{latex}{\out{\hypertarget{method-generateData}{}}}
\subsection{Method \code{generateData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFNetwork$generateData(
  numSamples = 1000,
  doPlot = FALSE,
  errDist = "normal",
  latIdx = c(2, 3)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numSamples}}{number of samples in the generated data frame (default = 1000)}

\item{\code{doPlot}}{whether or not to generate a network plot (default = FALSE)}

\item{\code{errDist}}{error distribution of data generated for the network, as in rmvDAG}

\item{\code{latIdx}}{indices of hubs (in decreasing order) to make latent; by default 2nd and 3rd largest.  These will be prefixed with "U_" (unobserved).  If none, set to NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SFNetwork$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
